rename the files and update the goal
