+++ simulation_multiprocess.py
@@ -19,7 +19,7 @@ from matplotlib import pyplot as plt
 import multiprocessing as mp
 
 from quantum_circuit_builder import generate_full_circuit
-from simulation_common import (
+from qec_simulation_core import (
     ResultPoint,
     build_bb_code,
     build_decoder_from_circuit,
@@ -27,7 +27,7 @@ from simulation_common import (
     plot_points,
     save_summary_csv,
 )
-from shared_utilities import (
+from file_io_utils import (
     safe_json_loads,
     existing_counts_for_point,
     DEFAULT_MAX_SHOTS,
